+++ src/buganize/__main__.py
@@ -1,3 +1,0 @@
-from buganize.cli.main import start
-
-start()